[Refactor] MeetingRepository의 groupBy 쿼리를 애플리케이션 레벨로 변경 (#45)

+++ src/main/kotlin/com/ringgo/domain/meeting/dto/MeetingWithMemberCount.kt
@@ -10,7 +10,7 @@ data class MeetingWithMemberCount(
     val name: String,
     val icon: String,
     val status: MeetingStatus,
-    val memberCount: Int,
+    val memberCount: Long,
     val creator: User,
     val createdAt: LocalDateTime
 )
